refactor: handle multiple line logs
